atlasexec/migrate-lint: added --context and handle exit status code
This change adds --context to MigrateLint.
It's only useful in combination with Web=true, which exposed a new
edge case where Atlas returns status code 1 when linting fails.

Indication of linting failure can be observed using the summary
when using MigrateLint, but when using Web=true, there is no
summary, only a URL.

So this PR also exposes the status code of Migrate Lint when
runnning from MigrateLintError